after review; auth_test.go rename to getapikey_test.go
